:bug: (aps/actual): fix actual image for Yarn v4
Since Actual migrate from YARN v3 to v4, we can't use yarn-audit-fix
anymore. Futhermore, it's not a good practice to overwrite dependencies
because it can break something with any warning and I didn't test
properly the project to validate the new build.

Instead of relying on "audit fix" to fixup these issues, I prefer to
rely on the CI that will fail if a critical CVE was found.

I also fixed some APK packages in order to make the rebuild more
reproductible.p